conform to the hidden assumptions of Docker
The Docker server (the thing that talks to the registry during
`docker pull/push`) queries GET <hostname>/v2/ and then makes the
following assumptions:

- For any endpoint below <hostname>, the `realm` and `service` are
  always those that are presented in the auth challenge for the /v2/
  endpoint.
- The `scope` is always in the concrete format that the Docker Registry
  generates.

None of this is mandated by the registry API specs as far as I can tell.
But of course, following the spec to the letter is not helpful when
`docker pull/push` ain't working.