include token scope in username for keystone auth
By tacking on an additional "project@domain" in the username, we have
all the information in username:password that we need to obtain a
scoped token, and therefore to check for the access level granted by it.

For the /v2/_catalog endpoint, auth will now add additional scopes that
encode which accounts may be listed.