initial implementation of quotas
We put a quota on the number of manifests to ensure that our Postgres
does not fill up uncontrolled. We may also add "number of repos" as a
quota later on, but for now we KISS since we expect to be able to model
the scaling of customer payload with the manifest count quota alone.