bypass keppel-registry for manifest storage, store in StorageDriver directly
I'm going to do the same for blobs next, thus removing the need for
orchestrating a keppel-registry fleet entirely. The registry fleet was a
good idea initially, but our requirements have grown and we now already
implement a large part of the registry API anyway. It therefore makes sense
to go the full way and remove the huge complexity introduced by the fleet
orchestration.