[ci skip] bypass keppel-registry entirely in Registry v2 API
The tests pass (except for a note that I left for myself, hence the "ci
skip"), but this Keppel is in a horribly broken state because the Keppel
v1 API still uses the OrchestrationDriver which is obviously nonsense.
I'm going to tackle that next. I'm also not satisfied with the code
structure in internal/api/registry/uploads.go, but I have a plan how to
make that part nicer. For now, I'm just glad to be at a point where it
makes sense to cut a commit; it's large enough as it is.