redesign replication to allow fully-consistent manifest references at all times
Replication used to work such that when the manifest is replicated in
the first step, the fact that all referenced blobs and manifests are
probably missing is just ignored since Keppel trusts that the upstream
account has a consistent state.

The problem with this approach is that I cannot properly garbage-collect
unreferenced blobs and manifests when the DB perspective on
blob/manifest references is not consistent to begin with.

The new method replicates manifests recursively to ensure consistent
manifest-manifest references, and creates "unbacked blobs" (with an
empty storage ID) in the DB to ensure consistent manifest-blob
references. The replication code path for blobs is triggered when the
GET handler encounters a blob with an empty storage ID.