Revert "do not count manifest syncs towards last_pulled_at"
This reverts commit 5bd4cc10c568f9cfb8652fedb6f0da3086b0668b.

This logic does not work in productive environments because of the bug
fixed by this commit's parent commit: In prod, a HEAD response does not
have a response body, so we cannot decode it into a RegistryV2Error.
This only accidentally worked during testing because the test HTTP
server did not strip response bodies for HEAD requests.