chunk large blobs during monolithic upload
This reuses and abstracts the existing chunking logic used for
replicating blobs.

The multi-part blob upload does not use chunking yet. I will do this in
the next commit, but this part will require some more refactoring since
the present chunking logic assumes that the Content-Length is known, but
streamed blob upload does not give us a Content-Length to work with.