add table `manifest_contents`
We had an outage because the Keystone DB pod was pulling a new image
from Keppel, which was trying to get the manifest from Swift, which was
trying to validate the Keppel service user token in Keystone, which was
down because its DB pod was restarting.

In all likelihood, this was never noticed before because Swift has some
caching for Keystone tokens, and this was the first time we had a
Keystone DB pod restart just as the Keppel service user token expired
in Swift's token cache.

To combat this, the new `manifest_contents` will keep manifests in the
Keppel DB, so that we don't need Swift for serving manifests.