try harder not to leak response bodies
I'm investigating a goroutine leak in keppel-janitor. I nuked a
longrunning janitor process to get goroutine backtraces, and got a
bunch of:

    goroutine 15765 [select, 2386 minutes]:
    net/http.(*persistConn).writeLoop(0xc0005cb9e0)
            /usr/local/go/src/net/http/transport.go:2382 +0xf7
    created by net/http.(*Transport).dialConn
            /usr/local/go/src/net/http/transport.go:1744 +0xc9c

    goroutine 110269 [select, 517 minutes]:
    net/http.(*persistConn).readLoop(0xc0004ebe60)
            /usr/local/go/src/net/http/transport.go:2203 +0x967
    created by net/http.(*Transport).dialConn
            /usr/local/go/src/net/http/transport.go:1743 +0xc77

So my suspicion is that response bodies do not get closed properly.
I'm honestly not sure if this bug even originates in my own code, but
properly closing response bodies is good practice anyway. We'll see if
it helps.