replace domain-mapping middleware with native domain remapping
The main motivation for this is that repository-level operations on the
domain-remapped APIs were not working with the standard Docker client:
When pulling from or pushing to a domain-remapped API, for instance:

  docker pull foo.registry.example.org/bar

The Docker client does not do what a sane client would do: attempt the
actual GET operation, get a 401 with auth challenge, and then do *that*
auth challenge. Instead, it looks at the auth challenge for `GET /v2/`
and then GUESSES the token scope based on the image reference. In our
example, it guesses a scope of `repository:bar:push` instead of what the
correct auth challenge would have requested, `repository:foo/bar:push`.

We have no choice but to play along with this nonsense. At the very
least, this would mean that the domain-remapping middleware now has to
parse, rewrite and reserialize both outgoing token challenges and
incoming token requests. And even then, there is the possibility that
some clients look at the issued tokens and demand to see exactly the
scopes that they requested, which would then also mean parsing,
rewriting and reissuing tokens in the middleware.

Clearly, this is insanity, so this commit brings domain-remapping into
the core as a first-class feature. This is one of these commits that are
unfortunately rather large since the effects of this change are fairly
widespread. On the bright side, the change is probably much smaller than
it would have been one or two months ago, since the recent auth refactor
has centralized most code that needs adjusting.

The major changes are as follows:

- auth.Service becomes auth.Audience, and now tracks both the
  anycast-ness of the request as well as its domain-remapped-ness.

- Tokens for domain-remapped APIs indicate their domain-remapped-ness in
  the audience and issuer fields. This is important because scopes like
  `repository:foo/bar:pull` refer to different repositories (and
  possibly even different accounts!) on the main API vs. on a
  domain-remapped API.

- As a consequence, auth.Scope gains a new method, ParseRepositoryScope,
  in which the interpretation of regular vs. domain-remapped repository
  references is centralized.

- When reverse-proxying anycast requests into a foreign Keppel API, we
  need to be careful to choose the domain-remapped foreign API if
  necessary to ensure that the user's token gets accepted over there.