guard against nil pointer dereference
This is to fix a crash in prod, although I'm not sure why the crash even
happens. All the history entries have non-nil "created" values, the only
thing out of the ordinary is that the first layer has the Unix epoch as
its creation timestamp:

    { "created": "1970-01-01T00:00:00Z", ... }

Maybe that gets misinterpreted as zero and hence nil?