janitor: add jitter to task schedules
This came out of an investigation into our vulnerability check
performance. The queue length metric is very jumpy: Sometimes 1000
images become due for a vulnerability check at the same time, and then
eventually they are all checked in rapid succession and the queue length
drops by 1000 again. That makes the queue length metric not very useful
as an overall measurement of how far we're lagging behind in
vulnerability checks.

With this change, once a task is completed, we schedule the next task
with a certain random variation. For example, a vulnerability check is
supposed to be executed hourly. With this change, we're setting the next
scheduled check not at exactly 60 minutes from now, but at somewhere
between 54-66 minutes from now. This should smoothen out the workload
and thus make the queue length metric steadier and thus more useful.

I've also applied this change to all other tasks that are driven by a
"next_XXX_at" schedule. We have some tasks (namely, blob validation and
manifest validation) that are driven in the inverse way (by "fixed time
since last check"); I did not adjust these since it is not very easy to
add jitter here. Longterm, I would like to move these over into the
forward-looking schedule model instead, but that's a change for another
day.