refactor Accept handler handling to reduce cognitive complexity
The special case rules get moved into the ParsedManifest interface.
According to gocognit, this reduces the complexity of
handleGetOrHeadManifest() from 100 to 66. I also used the opportunity to
turn an incorrect 404 response into the more appropriate 406 (Not
Acceptable).

A change was needed in the RepoClient since my understanding of the
Accept header was actually incorrect. It's not allowed to have multiple
Accept headers. Instead, all acceptable MIME types, must be listed in a
single header.