cmd/validate: allow multiple args, cache validation results
We already had a cache for the scope of a single ValidateManifest() and
its recursive subcalls, but that was not really helpful. We only stand
to gain when reusing the validation results for common base layers
shared by multiple images.

```
$ ./build/keppel validate alpine:latest alpine:3.17 --platform-filter '[{"os":"linux","architecture":"amd64"}]'
2023/03/17 16:38:57 INFO: interpreting alpine:latest as docker-pullable://registry-1.docker.io/alpine:latest
2023/03/17 16:39:23 INFO: manifest sha256:ff6bdca1701f3a8a67e328815ff2346b0e4067d32ec36b7992c1fdc001dc8517 looks good
2023/03/17 16:39:23 INFO:   manifest sha256:e2e16842c9b54d985bf1ef9242a313f36b856181f188de21313820e177002501 looks good
2023/03/17 16:39:29 INFO:     blob     sha256:b2aa39c304c27b96c1fef0c06bee651ac9241d49c4fe34381cab8453f9a89c7d looks good
2023/03/17 16:39:50 INFO:     blob     sha256:63b65145d645c1250c391b2d16ebe53b3747c295ca8ba2fcb6b0cf064a4dc21c looks good
2023/03/17 16:39:50 INFO: interpreting alpine:3.17 as docker-pullable://registry-1.docker.io/alpine:3.17
2023/03/17 16:39:53 INFO: manifest sha256:ff6bdca1701f3a8a67e328815ff2346b0e4067d32ec36b7992c1fdc001dc8517 looks good
2023/03/17 16:39:53 INFO:   manifest sha256:e2e16842c9b54d985bf1ef9242a313f36b856181f188de21313820e177002501 looks good (cached result)
```

This change is interesting when `keppel validate` is used to force
replication of lots of images at once, like in the Kubernikus pipeline.

While adding in the long-lived cache in the new ValidationSession
struct, I also refactored the validation functions in internal/client/
to be more internally consistent: Both ValidateManifest and
ValidateBlobContents now have a public facade and a private do...()
function with similar interfaces and behavior with regards to logging
and caching.

I also changed the cache to only remember positive results. In all
current usecases (cmd/anycastmonitor, cmd/healthmonitor, cmd/validate),
negative results lead to immediate failure anyway, so we would never
even attempt to hit the cache again for an object that had a negative
cache result. I'm considering to add a retry loop for intermittent
validation errors (esp. "currently replicating on a different worker"),
and that would not play well with caching of negative validation
results.