limit Clair interactions to a max of 10 minutes per task
We have some janitor tasks, probably vulnerability checks, that are
stuck for hours. All we can see is that the DB connection is in state
"idle in transaction", so my best guess is that the requests to Clair do
not complete in a timely manner. Adding a timeout, however generous,
should at least allow us to see in the log which operation is so slow
that it gets timed out.