test: memoize generation of deterministic test layers
This shaves about 2 seconds off a full test run. Not a lot, but quite
decent given the small change. We could earn another slightly bigger
improvement by replacing bcrypt for the peer passwords with a weaker and
faster hash during unit tests, but that would be a more invasive change.