add new accounts.rbac_policies_json field
There is not really any advantage to having the RBAC policies in their
own table, so I want to move them into the accounts table, similar to
what we already have for GC policies and security scan policies.

This first phase just adds the new field and makes the Keppel API
populate it whenever it touches the rbac_policies table. Once we have
this rolled out and the new field is filled on all accounts, we will
move into the next phase where the old table is removed and all usages
are cut over to the new table.