replace `rbac_policies` table with `accounts.rbac_policies_json`
This simplifies a lot of the CRUD for accounts and RBAC policies, and
reduces the number of SQL queries for each auth by one, at the expense
of making `accounts` records a bit larger.

I have verified that the logic from the previous PR has migrated all
existing policies into the new DB column, so it is safe to drop the
old table.