replace `validated_at` fields with `next_validation_at` fields
This is one of the last holdouts of a task clock field that refers to
the last run instead of the next run, so this is a neat little win for
the consistency department.

The main impetus of this change, though, is that the discovering queries
for these tasks rank #3 and #8 on our list of most expensive statements
(i.e. the pg_stat_statements view, ranked by total_exec_time).
The rephrasing from `validated_at` into `next_validation_at` simplifies
these queries considerably.

The actual implementation is sort of a case of "it can't be this hard...
oh no". Most of the diff is in test fixtures, though. The only
significant reworking that was required is in package processor, where
validateAndStoreManifestCommon() needed to be extended to explicitly be
told if the manifest is new or existing, since checking
`PushedAt == ValidatedAt` was not feasible anymore.