rewrite FindOrCreateRepository to prefer a plain SELECT
The INSERT-first approach with ON CONFLICT DO NOTHING is often neat, but
in this case, it does not even save queries.

And what's worse, the INSERT always advances repo_id_seq even if no
record ends up being inserted, so this always incurs a write in the DB
even if we only read an existing repo.