drivers/openstack: use more compact serialization for Keystone tokens
Since the new format should have all redundancies eliminated from it,
I figure that the roundtrip through gzip and base64 is not going to
save much space (or could actually waste bytes if gzip cannot save more
than the 33% that base64 adds). In any case, it's more computationally
expensive, so I'm getting rid of it.