api/registry: respond 401 with challenge when anonymous first pull is rejected
The existing behavior breaks `helm pull`, which (as intended by the API spec)
tries to GET a manifest to be presented with a 401 error and
Www-Authenticate challenge if necessary. We did not do this before,
because _technically_ anonymous users could pull manifests _if_ they
were replicated. But then they were not, and we got stuck.

With this, Helm and friends should try to authenticate and then retry
the GET as an authenticated user.