store Trivy reports in StorageDriver
The final intent here is that, when a user asks for a Trivy report on
the API, we can skip calling out to Trivy if we already have a report
cached from a recent security status check.

The API change is not in here because the PR is already quite big as it
is: What this commit does is to

- extend StorageDriver methods to read/write/list Trivy reports,
- adjust the storage sweep job to cleanup unexpected Trivy reports
  (in the same way as it already does for blobs and manifests),
- extend the security status check job to store Trivy reports, and
- extend Processor.DeleteManifest() to cleanup stored Trivy reports.

The DB schema is extended to track Trivy reports in storage, as well as
unexpected Trivy reports that are in the process of being GC'd by the
storage sweep.