fix DB migration
This migration does not work as-is. In real deployments, we have plenty
of images in "Rotten" that have `next_check_at` because that field is
NOT NULL before this migration.

In order to allow the CHECK constraint to be added, we need to change
those to a different vuln_status. "Pending" is a sensible choice.